requirements.txt
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.0.1
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
simple-websocket==0.5.1
six==1.16.0
smbus2==0.4.1
unifr-api-epuck==1.1.4
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.0
wsproto==1.1.0
zipp==3.7.0
